requirements.txt
# Web Framework & Server
fastapi
uvicorn[standard]
pydantic
python-dotenv

# Data Processing & Analysis
pandas
numpy

# Machine Learning & NLP (For Fraud Detection, SMS Analysis, & Call Warning)
scikit-learn
transformers
torch
spacy
nltk
regex

# Agentic AI — LangGraph for orchestration, LangChain for tool integration
langgraph
langchain-core
langchain-google-genai
langchain

# LLM SDK (Direct Gemini access for non-agentic calls)
google-genai

# Database & ORM (For India-Specific Scam Library & Risk Score History)
sqlalchemy
alembic

# Utilities
requests
beautifulsoup4
